impr - move token copying functions in own file
